+++ packages/toolkit/config/__tests__/webpack-fast-refresh.js
@@ -1,0 +1,66 @@
+import { getBuildFiles as getBuildFilesMock } from '../../utils/config';
+import { hasProjectFile as hasProjectFileMock } from '../../utils/file';
+import { getPackage as getPackageMock } from '../../utils/package';
+import webpackSerializer from '../../test-utils/webpack-serializer';
+
+jest.mock('../../utils/package', () => {
+	const module = jest.requireActual('../../utils/package');
+
+	jest.spyOn(module, 'getPackage');
+
+	return module;
+});
+
+jest.mock('../../utils/config', () => {
+	const module = jest.requireActual('../../utils/config');
+
+	jest.spyOn(module, 'getBuildFiles');
+
+	return module;
+});
+
+jest.mock('../../utils/file', () => {
+	const module = jest.requireActual('../../utils/file');
+
+	jest.spyOn(module, 'hasProjectFile');
+
+	return module;
+});
+
+describe('webpack.config.js', () => {
+	beforeAll(() => {
+		expect.addSnapshotSerializer(webpackSerializer);
+	});
+
+	it('returns proper webpack fast refresh configs for project configs', () => {
+		let webpackConfig;
+		jest.isolateModules(() => {
+			// eslint-disable-next-line global-require
+			webpackConfig = require('../webpack-fast-refresh.config');
+		});
+
+		expect(webpackConfig).toMatchSnapshot();
+	});
+
+	it('includes react-webpack-fast-refresh with the --hot option', () => {
+		process.argv.push('--hot');
+		process.env.NODE_ENV = 'development';
+		hasProjectFileMock.mockReturnValue(true);
+		const entryBuildFiles = {
+			entry1: 'entry1.js',
+		};
+		getBuildFilesMock.mockReturnValue(entryBuildFiles);
+		getPackageMock.mockReturnValue({
+			'10up-toolkit': {
+				entry: entryBuildFiles,
+			},
+		});
+		let webpackConfig;
+		jest.isolateModules(() => {
+			// eslint-disable-next-line global-require
+			webpackConfig = require('../webpack.config');
+		});
+
+		expect(webpackConfig).toMatchSnapshot();
+	});
+});
